Refactor computestorage hierarchy to group them together.
