+++ src/ispresent.rs
@@ -341,51 +341,3 @@ pub mod hypervdevicevirtualization {
         unsafe { IsHdvDeliverGuestInterruptPresent() != 0 }
     }
 }
-
-pub mod computestorage {
-    use crate::computestorage_bindings::*;
-
-    pub fn import_layer_is_present() -> bool {
-        unsafe { IsHcsImportLayerPresent() != 0 }
-    }
-
-    pub fn export_layer_is_present() -> bool {
-        unsafe { IsHcsExportLayerPresent() != 0 }
-    }
-
-    pub fn export_legacy_writable_layer_is_present() -> bool {
-        unsafe { IsHcsExportLegacyWritableLayerPresent() != 0 }
-    }
-
-    pub fn destroy_layer_is_present() -> bool {
-        unsafe { IsHcsDestroyLayerPresent() != 0 }
-    }
-
-    pub fn setup_base_os_layer_is_present() -> bool {
-        unsafe { IsHcsSetupBaseOSLayerPresent() != 0 }
-    }
-
-    pub fn initialize_writable_layer_is_present() -> bool {
-        unsafe { IsHcsInitializeWritableLayerPresent() != 0 }
-    }
-
-    pub fn initialize_legacy_writable_layer_is_present() -> bool {
-        unsafe { IsHcsInitializeLegacyWritableLayerPresent() != 0 }
-    }
-
-    pub fn attach_layer_storage_filter_is_present() -> bool {
-        unsafe { IsHcsAttachLayerStorageFilterPresent() != 0 }
-    }
-
-    pub fn detach_layer_storage_filter_is_present() -> bool {
-        unsafe { IsHcsDetachLayerStorageFilterPresent() != 0 }
-    }
-
-    pub fn format_writable_layer_vhd_is_present() -> bool {
-        unsafe { IsHcsFormatWritableLayerVhdPresent() != 0 }
-    }
-
-    pub fn get_layer_vhd_mount_path_is_present() -> bool {
-        unsafe { IsHcsGetLayerVhdMountPathPresent() != 0 }
-    }
-}
